refactor: 폴더명 변경으로 인하여 import 구문 변경
